requirements.txt
requests>=2.32.0
feedparser>=6.0.11
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
apscheduler>=3.10.4
pytz>=2024.1
tenacity>=8.2.3
